pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "environs>=14.5.0",
    "fastapi[standard]>=0.123.5",
    "pydantic-settings>=2.12.0",
    "stp-database",
]

[tool.uv.sources]
stp-database = { git = "https://github.com/STP-Team/stp-database" }

[tool.ruff.lint]
select = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"
